requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.1.0
blend_modes==2.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
docker-pycreds==0.4.0
dogma.py==0.0.1
einops==0.8.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.12.0
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.3
numpy==2.0.2
opencv-python==4.11.0.86
OpenEXR==3.3.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.7
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.25.2
scipy==1.15.2
sentry-sdk==2.24.0
setproctitle==1.3.5
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tifffile==2025.3.13
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.6.1
torchtyping==0.1.5
torchvision==0.21.0+cu118
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.8
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
